fix: page store path for SvelteKit breaking changes (#1615)
# Motivation

Last version of SvelteKit (bumped in #1612) contains another breaking changes. The route id of the `$app/stores.page` store now also prefix route group with `/`.

Used to be: `(app)/neuron`
Is now:  `/(app)/neuron`

This breaks our route matcher `pathForRouteId`
# Changes

- the related constant for the group `(app)` should be adapted to `/(app)`

# Tests

We cannot detect this with unit test because we have to mock the SvelteKit store. We could have detect it with e2e tests. I detected it on the "Sns neuron detail" page because the Skeleton did not disappeared.