feat: keep track of http agent (#1600)
# Motivation

This PR was original meant to introduce the `syncTime` feature of agent-js. To do so, we would have need to keep track - keep in memory - the `HttpAgent` that have been used because the `syncTime` function has to be called after the agent is created and because we would have like to call it only once.

Ultimately, or currently, it seems we do not need the `syncTime` feature anymore - because the network seems to have been fixed. Therefore we could have close the PR.

However, keeping track of the http agent seems to improve the performance of the app. Indeed, not re-creating the http agent on each API call seems to improve the overall behavior of the app.

That's why I suggest to go on with this PR.

# Follow-up

Previous PRs:

- Introduced https://github.com/dfinity/nns-dapp/pull/1584
- Reverted https://github.com/dfinity/nns-dapp/pull/1601

# Changes

- move `createAgent` in an api
- add a global variable that keeps track of the agent related to their principal - i.e. one anonymous and one for the user
- reset variable on sign-out