Test local build (#1681)
# Motivation
We wish to update e2e tests to make them more realistic. This PR focuses
on the internet identity.
- Current e2e tests use the version of II that is bundled with dfx. This
is not the version in production, not the version used on testnets and
not the version specified in the local dfx.json

# Changes
- Install the version of II specified in the local `dfx.json`.
- Update the e2e tests to match the newer II.
- Deploy II and the nns-dapp with `dfx nns install`. This ensures that
the canister IDs tested in the e2e tests match those depolyed by dfx and
we don't need a separate build for dfx. It also means that when `dfx nns
install` deploys the front end canisters to the production canister IDs,
our e2e tests can do the same with minimal effort.
- Note: Having fewer different ways of deploying should increase the
fidelity of testing and reduce the effort needed to maintain
deployments, however developers are still free to deploy canisters in
other ways.
- Use the upstream name for `internet_identity_dev.wasm` as normally
`internet_identity.wasm` is a production wasm and this can be confusing.
- Test that the II and nns-dapp wasms installed for e2e tests are the
ones we expect to be present.