feat: A minimal launchpad e2e test (#1726)
# Motivation
An e2e test is needed for load generation. It doesn't need to actually
test anything, it just needs to load the launchpad to stress the
network.

Proper e2e tests are in the pipeline for the launchpad, however stress
testing doesn't care about delicacies such as using tids to click
buttons. So the stress-testing e2e test can be used now without waiting
for tids et al.

# Changes
- Add an e2e test that loads the launchpad, first logged out then logged
in.
- Added an env var `WDIO_VIEW=<headless|yesplease>` so that the browser
can be watched without code changes.

# Tests
- I have run this e2e test locally and it will be run in CI:
  `WDIO_VIEW=yesplease npm run test -- --spec specs/launchpad.e2e.ts`

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>