GIX-1212: Add Proposal Status filter (#1723)
# Motivation

User can filter proposal by decision status.

IMPORTANT: This PR includes setup of filters, button, modal and store.
Not yet updating the proposals list.

# Changes

Main changes:
* SnsProposalsFilters.svelte will include all filters buttons and render
each modal.
* SnsFilterStatusModal.svelte shows filters and performs the actions in
the store.
* sns-filters.store manages filters for SNS projects.
* loadSnsFilters in sns-filters.services.

The idea to store the Filter type in the store is to use the same
concept for all SNS filters. Recall that SNS topics is not an enum,
instead is the list of nervous functions. Which have a name, id and
description.

I believe this is a good idea also for NNS. (for later) This way the UI
doesn't care or know where the filters come from. Whether an enum or
from a backend. The UI components are more reusable. We just need to add
the logic of filling up the store with the data.

Minor changes:
* New UI component FiltersWrapper.svelte to be reused in SNS and NNS
filters.
* ProposalsFilters.svelte to NnsProposalsFilters.svelte

# Tests

* Test new modal.
* Test new store.
* Test new service.
* Fix broken tests.