Asset reproducibility (#1734)
# Motivation
We want to progress:


![image](https://user-images.githubusercontent.com/5982633/213559458-e146d422-453d-4677-b6c0-7b4421853dbc.png)

This PR introduces a test but does not (yet) block PRs. The important
thing now is to decide what inputs should fully define the output of the
frontend build. The proposal here is that the output may depend on:

- code commit, including our own code and the
`frontend/packages-lock.json` file.
- node version
- npm version
- tar
- xz

but should NOT depend on:
- build environment including:
  - time
  - operating system
  - docker
  - CPU

# Changes
- Extract the frontend building code from the build script.
- Run the frontend building code on:
  - different OSs (Ubuntu and Mac)
  - different versions of the OSs
- at different times (now and with an artificial time shift into the
future)

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>