Fix CI required checks (#1759)
# Motivation
A PR was merged even though the corresponding test failed. If a
dependent test failed, the final check is skipped and historically that
would prevent the PR from merging. It seems that "needs" is now treated
as a conditional, so "skipped" is treated as success.
https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/defining-the-mergeability-of-pull-requests/troubleshooting-required-status-checks#handling-skipped-but-required-checks
This is not desired

# Changes
- Always run the final check, to see whether required GitHub jobs have
succeeded.

# Tests
- I made a toy job that always fails, set it as a dependency and
verified that the final check was run, not skipped, and failed.
- I repeated the same with the toy job succeeding, and the final check
similarly succeeded.

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>