feat: token store (#1804)
# Motivation

To integrate ckBTC transactions we need the requested ledger fee.

Currently in Sns, such fees are fetched with a dedicated `transferFee` which generate unecessary request on the backend as the fees are already provided and fetched with the `metadata` call.

Ultimately such Sns requests will be deprecated but, we need to integrate ckBTC now. In addition, as we are concern about performances, we want to avoid to implement uncessery calls for ckBTC as well.

That's why we should not query `metadata` and `balance` to load and maps our "Account" construct but, perform only what's necessary.

Therefore, this PR introduces a new `tokensStore`, adds few `TODOs` and set `deprecation` information which shall be implemented in upcoming PRs.

# Breaking Changes

As for `name` and `symbol`, `fee` becomes a mandatory field of the Icrc metadata. Sns-1 and ckBTC both provides such metadata information and FI team confirmed that an Icrc ledger metadata fee is in sync with the exposed endpoint that returns the fee.

# Changes

- mark `account.balance` as **deprecated** and to be replaced with `account.balanceE8s`
- mark `metadata` function of `icrc-ledger.api` as **deprecated**
- mark `transaction-fees.store` as **deprecated**
- mark `snsTokenSymbolSelectedStore` as **deprecated**
- introduces new `tokenStore` and related derived store to support universes with Nns
- add `getIcrcToken` to `icrc-ledger` api and usage in `sns/ckBTC-api`
- load token information in services
  - sns uncertified query account balance
  - icrc uncertified query account balance
  - `syncSnsAccounts`
  - `syncIcrcAccounts`
  - `loadSnsProjects`
- replace `loadIcrcAccounts` in cmp with `syncIcrcAccounts`
- use new `universesTokenStore` in `<UniverseAccountsBalance />`
- refactor `SnsTokenMetadata` to `IcrcTokenMetadata`
- add `fee` to `IcrcTokenMetadata`
- move `mapOptionalToken` to `icrc-tokens.utils`