Handle missing logo (#1850)
# Motivation
The sns aggregator provides a logo URL for every SNS. This can be used
directly in HTML without a custom API call. The code that generates this
URL does not, however, handle the case where the logo is not provided.

# Changes
- If the logo is not provided, provide an empty file. Note: It can be
debated whether we should return a 404, an empty file or some default
image. The priority now is to make the canister not break. _maybe_ the
404 is the better option, as it allows fallback images, although it will
print error messages in the console and red error messages shout "page
is broken" which isn't nice either. Maybe certify a pretty HTML message
explaining that the SNS has no logo along with the 404 so that there are
no error messages in the console and users have a good experience. Once
we have agreed, this would be a good follow-up task. It is not a trivial
matter of just changing the default text so would better be done in a
separate PR.