feat: remove lazy load route (#1875)
# Motivation

We noticed issue with route not being destroy. Most recent issue was solve with a workaround in PR #1834.

While I am not sure what's the root cause of the issue, I tend to think that it might be linked with the fact that we lazy loaded routes with a `RouteModule` component.

Today I had a try and removed this feature.

It seems it had not a bad impact on JS chunks loading. Without it we roughly load the same number of chunks, on loading page it's even the same number of chunks.

And in addition, according my test, `onDestroy` of Wallet pages were always called.

Therefore this PR removes `RouteModule` component.

# Changes

- remove `RouteModule` component
- inegrate use of component in `+page`
- tweak `NnsProposals` component to make pre-rendering pass (list is not pre-rendered anyway)
- tweak few `back` function to make pre-rendering pass