Enforce npm and nodejs version range (#1906)
# Motivation
Using an incompatible version of nodejs breaks builds, formatting and
lockfiles.

# Changes
- Cause node to fail if an incompatible version of nodejs or npm is
used.
- Note: I have specified this individually in the .npmrc of each node
project. We may want to make a global entry, but must the n also
remember to include it in docker builds that take just one directory and
named other files.
- Specify the version of nodejs in CI.
- Note: It would be nice to be able to read the node version from the
package.json. Ideally that should be done in both CI and in the
dockerfile. Should be possible, but in another PR.

# Tests
See CI