Fail tests if there are console.log/debug/error (#1930)
# Motivation

Getting log output while running `npm run test` is quite distracting and
usually not intended.
If the logging is expected, the test should really mock it out and
expect the behavior which means the logs don't appear in the output.

# Changes

1. Fail a test if console.log/debug/error happens during the test.
2. Only fail after the test finishes to allow seeing any other results
from the test.
3. For debugging allow disabling the functionality by calling
`allowLoggingInOneTestForDebugging()` from the test or beforeEach().
4. Edit existing tests that silence the console in a way that's broken
by this PR to keep them silencing the console.

# Tests

Manually checked that adding console.log to a test caused it to fail and
that disabling worked.