Reduce inputs to build steps (#1933)
# Motivation
The build steps for the nns-dapp frontend and backend currently run
./config.sh. For config.sh to work properly for environments other than
hard-wired entries such as mainnet, it needs access to the environment
such as `.dfx/`. The environment contains a huge amount of information
that is not relevant and that makes it harder to reason about the build
and breaks caches. If we put this "config extraction" into a separate
step, the inputs to the build can stay small.

# Changes
- Add a separate docker stage to get build configuration
- Move ./config.sh out of the frontend build command.
- Minimize inputs to the build commands, e.g. removing env vars.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>