fix: voting UI and cache (#1975)
# Motivation

The introduction of the neurons cache leads to an issue when the user
vote. Because the list of neurons is not refreshed with the very last
ballots, the voting actions ("buttons yes no") remains notably active
after vote.

There is also a minor UI quirks, a line is positioned absolutely to the
modal instead of parent.

Also we were not able to create dummy proposals anymore because the URL
of the proposal are white listed now.