Fast updates during a sale (#1984)
# Motivation
During a sale, the amount pledged by investors can change fairly
quickly, so it behoves us to update the cache fairly frequently.

# Changes
* Add a "Fast" scheduler
* The scheduler updates the swap data frequently while there is an open
sale.
* If there are no open sales but the date of a future sale is known, the
scheduler sets a timer for the next sale and goes to sleep.
* If there are no current or pending sales, the scheduler goes to sleep
and relies on the normal updates to detect any new pending sales when
they appear.

# Tests
* I have tested this manually on small09:
* aggregator:
https://tqtu6-byaaa-aaaaa-aaana-cai.small09.testnet.dfinity.network/
* nns-dapp:
https://tcvdh-niaaa-aaaaa-aaaoa-cai.small09.dfinity.network/project/?project=x2dwq-7aaaa-aaaaa-aaaxq-cai
* Watch the logs. If all SNSs are bought out, the fast updates cease. If
a new SNS is created, fast updates restart.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>